Title: 특정 문자로 시작하는 문자열, Time: 115ms, Memory: 10MB, Status: Passed - Codetree

+++ 250331/특정 문자로 시작하는 문자열/strings-that-start-with-a-specific-character.java
@@ -24,7 +24,7 @@ public class Main {
             }
         }
         
-        double avg = cnt/n;
+        double avg = (double)cnt/n;
 
         System.out.printf("%d %.2f",n,avg);
     }
